refactor(Score) : 원시값인 상수를 포장 BLACKJACK과 TEN을 Score 객체로 변경

+++ src/main/java/blackjack/controller/GameController.java
@@ -69,7 +69,7 @@ public class GameController {
         players.forEach(player -> askAddCardOrPass(player, deck));
         while (dealer.canAddCard()) {
             dealer.addCard(deck.drawCard());
-            OutputView.showDealerAddCard(Dealer.TURN_OVER_COUNT);
+            OutputView.showDealerAddCard(Dealer.TURN_OVER_COUNT.toInt());
         }
         dealer.stayIfNotFinished();
     }
